requirements.txt
kafka
kafka-python-ng
pyspark
setuptools
requests
datetime
python-dotenv
plotly
pandas
flask